+++ internal/compiler/llr/optim_passes/inline_expressions.rs
@@ -47,6 +47,7 @@ fn expression_cost(exp: &Expression, ctx: &EvaluationContext) -> isize {
         Expression::Struct { .. } => 1,
         Expression::EasingCurve(_) => 1,
         Expression::LinearGradient { .. } => ALLOC_COST,
+        Expression::RadialGradient { .. } => ALLOC_COST,
         Expression::EnumerationValue(_) => 0,
         Expression::ReturnStatement(_) => 1,
         Expression::LayoutCacheAccess { .. } => PROPERTY_ACCESS_COST,