+++ internal/core/graphics/brush.rs
@@ -26,6 +26,9 @@ pub enum Brush {
     /// The linear gradient variant of a brush describes the gradient stops for a fill
     /// where all color stops are along a line that's rotated by the specified angle.
     LinearGradient(LinearGradientBrush),
+    /// The radial gradient variant of a brush describes a circle variant centered
+    /// in the middle
+    RadialGradient(RadialGradientBrush),
 }
 
 /// Construct a brush with transparent color
@@ -44,6 +47,9 @@ impl Brush {
             Brush::LinearGradient(gradient) => {
                 gradient.stops().next().map(|stop| stop.color).unwrap_or_default()
             }
+            Brush::RadialGradient(gradient) => {
+                gradient.stops().next().map(|stop| stop.color).unwrap_or_default()
+            }
         }
     }
 
@@ -59,6 +65,7 @@ impl Brush {
         match self {
             Brush::SolidColor(c) => c.alpha() == 0,
             Brush::LinearGradient(_) => false,
+            Brush::RadialGradient(_) => false,
         }
     }
 
@@ -76,6 +83,11 @@ impl Brush {
                     position: s.position,
                 }),
             )),
+            Brush::RadialGradient(g) => {
+                Brush::RadialGradient(RadialGradientBrush::new_circle(g.stops().map(|s| {
+                    GradientStop { color: s.color.brighter(factor), position: s.position }
+                })))
+            }
         }
     }
 
@@ -88,6 +100,10 @@ impl Brush {
             Brush::SolidColor(c) => Brush::SolidColor(c.darker(factor)),
             Brush::LinearGradient(g) => Brush::LinearGradient(LinearGradientBrush::new(
                 g.angle(),
+                g.stops()
+                    .map(|s| GradientStop { color: s.color.darker(factor), position: s.position }),
+            )),
+            Brush::RadialGradient(g) => Brush::RadialGradient(RadialGradientBrush::new_circle(
                 g.stops()
                     .map(|s| GradientStop { color: s.color.darker(factor), position: s.position }),
             )),
@@ -124,6 +140,23 @@ impl LinearGradientBrush {
     pub fn stops(&self) -> impl Iterator<Item = &GradientStop> {
         // skip the first fake stop that just contains the angle
         self.0.iter().skip(1)
+    }
+}
+
+/// The RadialGradientBrush describes a way of filling a shape with a circular gradient
+#[derive(Clone, PartialEq, Debug)]
+#[repr(transparent)]
+pub struct RadialGradientBrush(SharedVector<GradientStop>);
+
+impl RadialGradientBrush {
+    /// Creates a new circle radial gradient, centered in the middle and described
+    /// by the provided color stops.
+    pub fn new_circle(stops: impl IntoIterator<Item = GradientStop>) -> Self {
+        Self(stops.into_iter().collect())
+    }
+    /// Returns the color stops of the linear gradient.
+    pub fn stops(&self) -> impl Iterator<Item = &GradientStop> {
+        self.0.iter()
     }
 }
 
@@ -184,6 +217,43 @@ impl InterpolatedPropertyValue for Brush {
                         x.position = x.position.interpolate(&1.0, t);
                     }
                     Brush::LinearGradient(new_grad)
+                }
+            }
+            (Brush::SolidColor(col), Brush::RadialGradient(grad)) => {
+                let mut new_grad = grad.clone();
+                for x in new_grad.0.make_mut_slice().iter_mut() {
+                    x.color = col.interpolate(&x.color, t);
+                }
+                Brush::RadialGradient(new_grad)
+            }
+            (a @ Brush::RadialGradient(_), b @ Brush::SolidColor(_)) => {
+                Self::interpolate(b, a, 1. - t)
+            }
+            (Brush::RadialGradient(lhs), Brush::RadialGradient(rhs)) => {
+                if lhs.0.len() < rhs.0.len() {
+                    Self::interpolate(target_value, self, 1. - t)
+                } else {
+                    let mut new_grad = lhs.clone();
+                    let mut iter = new_grad.0.make_mut_slice().iter_mut();
+                    let mut rhs_stops = rhs.stops();
+                    while let (Some(s1), Some(s2)) = (iter.next(), rhs_stops.next()) {
+                        s1.color = s1.color.interpolate(&s2.color, t);
+                        s1.position = s1.position.interpolate(&s2.position, t);
+                    }
+                    for x in iter {
+                        x.position = x.position.interpolate(&1.0, t);
+                    }
+                    Brush::RadialGradient(new_grad)
+                }
+            }
+            (a @ Brush::LinearGradient(_), b @ Brush::RadialGradient(_))
+            | (a @ Brush::RadialGradient(_), b @ Brush::LinearGradient(_)) => {
+                // Just go to an intermediate color.
+                let color = Color::interpolate(&b.color(), &a.color(), t);
+                if t < 0.5 {
+                    Self::interpolate(a, &Brush::SolidColor(color), t * 2.)
+                } else {
+                    Self::interpolate(&Brush::SolidColor(color), b, (t - 0.5) * 2.)
                 }
             }
         }